+++ RushTablePOS/src/Clases/Autoridad.java
@@ -1,0 +1,4 @@
+package Clases;
+
+public class Autoridad {
+}
