+++ RushTablePOS/src/Controlador/main.java
@@ -1,0 +1,8 @@
+package Controlador;
+
+public class main {
+
+    public static void main(String [] args){
+        // new mainForm().setVisible(true);
+    }
+}
